+++ plugins/macroview/MacroDiff.cpp
@@ -1689,7 +1689,7 @@ void __fastcall TMacroView::InsertMacroToEditor(BOOL AllMacros)
     if(Size)
     {
       CurDir=new WCHAR[Size];
-      Info.Control(PANEL_ACTIVE,FCTL_GETCURRENTDIRECTORY,Size,reinterpret_cast<LONG_PTR>(CurDir));
+      Info.Control(PANEL_ACTIVE,FCTL_GETCURRENTDIRECTORY,(int)Size,reinterpret_cast<LONG_PTR>(CurDir));
     }
 #endif
     int Code=CreateProcess(NULL,Str,NULL,NULL,TRUE,0,NULL,CurDir,&si,&pi);
